backend - Pesquisa de usuario por id

+++ Backend/src/main/java/oi/github/maurodegaspari/meusservicos/repository/ClienteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import oi.github.maurodegaspari.meusservicos.model.ClienteModel;
 
-public interface ClienteRepository extends JpaRepository<ClienteModel, Integer>{
+public interface ClienteRepository extends JpaRepository<ClienteModel, Long>{
 
 }
